add dokka for store module
move key parser to tests
