+++ build.gradle
@@ -18,7 +18,7 @@ buildscript {
         classpath 'com.android.tools.build:gradle:3.6.0-beta04'
         classpath 'com.getkeepsafe.dexcount:dexcount-gradle-plugin:0.5.6'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$rootProject.ext.versions.kotlin"
-        classpath 'org.jetbrains.dokka:dokka-gradle-plugin:0.9.17'
+        classpath 'org.jetbrains.dokka:dokka-gradle-plugin:0.10.0'
     }
 }
 
@@ -70,6 +70,10 @@ project.ext.preDexLibs = !project.hasProperty('disablePreDex')
 
 subprojects {
     apply plugin: 'com.diffplug.gradle.spotless'
+    if (project.path == ":store") {
+        apply plugin: 'org.jetbrains.dokka'
+    }
+
     spotless {
         kotlin {
             target 'src/**/*.kt'
