+++ store/build.gradle
@@ -17,6 +17,8 @@ plugins {
 
 apply plugin: 'java'
 apply plugin: 'kotlin'
+apply plugin: 'org.jetbrains.dokka'
+
 
 group = GROUP
 version = VERSION_NAME
